refactor: Remove threaded compute function options (#20)
This was a half-baked part of the initial code that was never actually
used. Concurrency should almost certainly be handled above (e.g., dask)
or below (e.g., in geoarrow-c)...what ended up happening was that the
tasks took so little time to complete that it didn't end up helping
much.