perf: Remove validation from `wrap_array()` (#23)
This resulted in some unexpected performance characteristics and was not
all that useful (since you could still have geometrically invalid data
that happened to parse with `validate=True`). Because it was the
default, it was pretty easy to get terrible performance and not know why
(e.g., with `geoarrow.pyarrow.io.read_pyogrio_table()`).